+++ libraries/Bluefruit52Lib/src/BLEScanner.cpp
@@ -147,6 +147,8 @@ uint8_t BLEScanner::parseReportByType(const uint8_t* scandata, uint8_t scanlen, 
   uint8_t len = 0;
   uint8_t const* ptr = NULL;
 
+  if ( scanlen < 2 ) return 0;
+
   // len (1+data), type, data
   while ( scanlen )
   {
@@ -275,6 +277,16 @@ void BLEScanner::filterUuid(BLEUuid ble_uuid[], uint8_t count)
   _filter_uuid = new BLEUuid[count];
 
   for(uint8_t i=0; i<count; i++) _filter_uuid[i] = ble_uuid[i];
+}
+
+void BLEScanner::filterService(BLEService& svc)
+{
+  filterUuid(svc.uuid);
+}
+
+void BLEScanner::filterService(BLEClientService& cli)
+{
+  filterUuid(cli.uuid);
 }
 
 void BLEScanner::filterMSD(uint16_t manuf_id)